# brain-backend/api-ref/brain_openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Brain API",
        "version": "0.1.0"
    },
    "paths": {
...
        "/images": {
            "get": {
                "tags": [
                    "images"
                ],
                "summary": "Get All Images",
                "description": "Get all images list",
                "operationId": "get_all_images_images_get",
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "items": {
                                        "$ref": "#/components/schemas/Image"
                                    },
                                    "type": "array",
                                    "title": "Response Get All Images Images Get"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "OAuth2PasswordBearer": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "images"
                ],
                "summary": "Create Image",
                "description": "Create a new image",
                "operationId": "create_image_images_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ImageCreate"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Image"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "OAuth2PasswordBearer": []
                    }
                ]
            }
        },
        "/images/{image_id}": {
            "get": {
                "tags": [
                    "images"
                ],
                "summary": "Get Image",
                "description": "Get specific image by ID",
                "operationId": "get_image_images__image_id__get",
                "parameters": [
                    {
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Image Id"
                        },
                        "name": "image_id",
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Image"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "OAuth2PasswordBearer": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "images"
                ],
                "summary": "Update Image",
                "description": "Update image information by ID",
                "operationId": "update_image_images__image_id__put",
                "parameters": [
                    {
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Image Id"
                        },
                        "name": "image_id",
                        "in": "path"
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ImageUpdate"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Image"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "OAuth2PasswordBearer": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "images"
                ],
                "summary": "Delete Image",
                "description": "Delete image by ID",
                "operationId": "delete_image_images__image_id__delete",
                "parameters": [
                    {
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Image Id"
                        },
                        "name": "image_id",
                        "in": "path"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful Response"
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "OAuth2PasswordBearer": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
            "Image": {
                "properties": {
                    "id": {
                        "type": "string",
                        "title": "Id",
                        "description": "Unique identifier for the image"
                    },
                    "name": {
                        "type": "string",
                        "title": "Name",
                        "description": "Image name"
                    },
                    "ceph_location": {
                        "type": "string",
                        "title": "Ceph Location",
                        "description": "Ceph location in format pool/rbd"
                    },
                    "mon_host": {
                        "type": "string",
                        "format": "ipvanyaddress",
                        "title": "Mon Host",
                        "description": "Ceph monitor host IP address"
                    },
                    "description": {
                        "type": "string",
                        "title": "Description",
                        "description": "Image description"
                    }
                },
                "type": "object",
                "required": [
                    "id",
                    "name",
                    "ceph_location",
                    "mon_host"
                ],
                "title": "Image"
            },
            "ImageCreate": {
                "properties": {
                    "name": {
                        "type": "string",
                        "title": "Name",
                        "description": "Image name"
                    },
                    "ceph_location": {
                        "type": "string",
                        "title": "Ceph Location",
                        "description": "Ceph location in format pool/rbd"
                    },
                    "mon_host": {
                        "type": "string",
                        "format": "ipvanyaddress",
                        "title": "Mon Host",
                        "description": "Ceph monitor host IP address"
                    },
                    "description": {
                        "type": "string",
                        "title": "Description",
                        "description": "Image description"
                    }
                },
                "type": "object",
                "required": [
                    "name",
                    "ceph_location",
                    "mon_host"
                ],
                "title": "ImageCreate"
            },
            "ImageUpdate": {
                "properties": {
                    "name": {
                        "type": "string",
                        "title": "Name",
                        "description": "Image name"
                    },
                    "ceph_location": {
                        "type": "string",
                        "title": "Ceph Location",
                        "description": "Ceph location in format pool/rbd"
                    },
                    "mon_host": {
                        "type": "string",
                        "format": "ipvanyaddress",
                        "title": "Mon Host",
                        "description": "Ceph monitor host IP address"
                    },
                    "description": {
                        "type": "string",
                        "title": "Description",
                        "description": "Image description"
                    }
                },
                "type": "object",
                "title": "ImageUpdate"
            },
...
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": {
                        "type": "string",
                        "title": "Message"
                    },
                    "type": {
                        "type": "string",
                        "title": "Error Type"
                    }
                },
                "type": "object",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "title": "ValidationError"
            }
        },
        "securitySchemes": {
            "OAuth2PasswordBearer": {
                "type": "oauth2",
                "flows": {
                    "password": {
                        "scopes": {},
                        "tokenUrl": "login"
                    }
                }
            }
        }
    }
}